requirements.txt
wandb==0.17.7
transformers==4.36.1
tqdm==4.42.1
torchcontrib==0.0.2
torchstat==0.0.7
soundfile==0.12.1
scikit-learn==0.22.2.post1
opencv-python==4.4.0.40
opensmile==2.5.0
numpy==1.23.5
librosa==0.10.0
speechbrain==1.0.0